commiting modified changes

+++ TimeSheetManagement/account_project_service/src/main/java/com/feuji/accountprojectservice/bean/AccountBean.java
@@ -31,7 +31,7 @@ public class AccountBean {
 	private String zipcode;
 	private String country;
 	private Boolean isRed;
-	private Long accountStatus;
+	private Integer  accountStatus;
 	private String comments;
 	private Boolean isDeleted;
